+ Small change to allow compilation with gcc-4.6.2.

+++ src/cdi/test/DummyEoS.cc
@@ -83,7 +83,7 @@ std::vector< double > DummyEoS::getSpecificElectronInternalEnergy(
     const std::vector< double >& vdensity ) const
 {
     std::vector< double > seie( vdensity.size() );
-    for ( size_t i=0; i<vdensity.size(); ++i )
+    for ( unsigned i=0; i<vdensity.size(); ++i )
 	seie[i] = vtemperature[i] + 1000.0*vdensity[i];
     return seie;
 }
@@ -120,7 +120,7 @@ std::vector< double > DummyEoS::getElectronHeatCapacity(
     const std::vector< double >& vdensity ) const
 {
     std::vector< double > ehc( vdensity.size() );
-    for ( size_t i=0; i<vdensity.size(); ++i )
+    for ( unsigned i=0; i<vdensity.size(); ++i )
 	ehc[i] = vtemperature[i] + vdensity[i]/1000.0;
     return ehc;
 }
@@ -225,7 +225,7 @@ std::vector< double > DummyEoS::getNumFreeElectronsPerIon(
     const std::vector< double >& vdensity ) const
 {
     std::vector< double > nfepi( vdensity.size() );
-    for ( size_t i=0; i<vdensity.size(); ++i )
+    for ( unsigned i=0; i<vdensity.size(); ++i )
 	nfepi[i] = vtemperature[i]/100.0;
     return nfepi;
 }
@@ -263,7 +263,7 @@ std::vector< double > DummyEoS::getElectronThermalConductivity(
     const std::vector< double >& vdensity ) const 
 {
     std::vector< double > ebtc( vdensity.size() );
-    for ( size_t i=0; i<vdensity.size(); ++i )
+    for ( unsigned i=0; i<vdensity.size(); ++i )
 	ebtc[i] = 1000.0*vtemperature[i] + vdensity[i];
     return ebtc;
 }
